ref: Moved all script custom functions to its own file
